backend/main.py
"""
Multi-Agent AI System - FastAPI Backend
Main application entry point with API endpoints
"""

from fastapi import FastAPI, HTTPException
from fastapi.middleware.cors import CORSMiddleware
from fastapi.responses import JSONResponse
import uvicorn
from typing import Dict, Any
import os
from dotenv import load_dotenv

# Load environment variables
load_dotenv()

# Initialize FastAPI app
app = FastAPI(
    title="Multi-Agent AI System",
    description="A sophisticated multi-agent AI system with 7 specialized agents",
    version="1.0.0",
    docs_url="/docs",
    redoc_url="/redoc"
)

# CORS middleware configuration
app.add_middleware(
    CORSMiddleware,
    allow_origins=["http://localhost:3000", "http://127.0.0.1:3000"],  # Next.js dev server
    allow_credentials=True,
    allow_methods=["*"],
    allow_headers=["*"],
)

# Health check endpoint
@app.get("/health")
async def health_check():
    """Health check endpoint for monitoring"""
    return {
        "status": "healthy",
        "message": "Multi-Agent AI System is running",
        "version": "1.0.0"
    }

# Main query endpoint
@app.post("/query")
async def process_query(query_data: Dict[str, Any]):
    """
    Main endpoint for processing user queries through the multi-agent system
    
    Expected input:
    {
        "query": "What is the sentiment of recent AI news?",
        "user_id": "optional_user_id"
    }
    """
    try:
        query = query_data.get("query", "")
        user_id = query_data.get("user_id", "anonymous")
        
        if not query:
            raise HTTPException(status_code=400, detail="Query is required")
        
        # TODO: Implement agent orchestration logic
        # This will be implemented in Phase 2-3
        
        # Placeholder response
        response = {
            "status": "success",
            "query": query,
            "user_id": user_id,
            "message": "Query received - Agent orchestration will be implemented in Phase 2",
            "agents_used": [],
            "result": {
                "type": "placeholder",
                "data": "This is a placeholder response. The multi-agent system will be implemented in subsequent phases."
            }
        }
        
        return JSONResponse(content=response)
        
    except Exception as e:
        raise HTTPException(status_code=500, detail=f"Internal server error: {str(e)}")

# Agent status endpoint
@app.get("/agents/status")
async def get_agents_status():
    """Get status of all agents in the system"""
    agents = {
        "decision_agent": {"status": "not_implemented", "description": "Routes queries to appropriate agents"},
        "research_agent": {"status": "not_implemented", "description": "RAG with document retrieval"},
        "news_agent": {"status": "not_implemented", "description": "Fetches live news data"},
        "sentiment_agent": {"status": "not_implemented", "description": "Analyzes text sentiment"},
        "summarizer_agent": {"status": "not_implemented", "description": "Condenses outputs into insights"},
        "frontend_agent": {"status": "not_implemented", "description": "Formats results for UI"},
        "documentation_agent": {"status": "not_implemented", "description": "Auto-updates documentation"}
    }
    
    return {
        "status": "system_initialized",
        "agents": agents,
        "message": "All agents are ready for implementation in Phase 2"
    }

# System configuration endpoint
@app.get("/config")
async def get_system_config():
    """Get system configuration and environment info"""
    return {
        "environment": os.getenv("ENVIRONMENT", "development"),
        "openai_api_key_configured": bool(os.getenv("OPENAI_API_KEY")),
        "weaviate_url_configured": bool(os.getenv("WEAVIATE_URL")),
        "news_api_key_configured": bool(os.getenv("NEWS_API_KEY")),
        "cors_origins": ["http://localhost:3000", "http://127.0.0.1:3000"]
    }

if __name__ == "__main__":
    uvicorn.run(
        "main:app",
        host="0.0.0.0",
        port=8000,
        reload=True,
        log_level="info"
    )
