backend/requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# LangGraph and AI/ML
langgraph==0.0.20
langchain==0.1.0
langchain-openai==0.0.2
langchain-community==0.0.10

# Vector database
weaviate-client==3.25.3

# HTTP requests and APIs
httpx==0.25.2
requests==2.31.0

# Data processing
pandas==2.1.4
numpy==1.24.3

# Environment and configuration
python-dotenv==1.0.0

# CORS and middleware
python-cors==1.7.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0

# Optional: Database
# psycopg2-binary==2.9.9  # PostgreSQL
# sqlalchemy==2.0.23
